--- index.handlebars
@@ -1,2 +1,0 @@
-
-<script src="{{googleApi}}" async defer></script>
