+++ package.json
@@ -1,39 +1,19 @@
 {
-  "name": "gwg-women-techmakers",
-  "version": "0.0.1",
-  "description": "This is a Grow with Google project for an offline first app that provides information based on geolocation",
-  "main": "src/app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack-dev-server",
-    "build:prod": "webpack -p"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/gwg-women/gwg-women-techmakers.git"
-  },
-  "contributors": [
-    "Tanya Gupta (https://github.com/tanyagupta)",
-    "Khusbu Chandra (https://github.com/khusbuchandra)",
-    "Amy Carney (https://github.com/digilou)"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/gwg-women/gwg-women-techmakers/issues"
-  },
-  "homepage": "https://github.com/gwg-women/gwg-women-techmakers#readme",
-  "devDependencies": {
-    "babel-preset-env": "^1.6.1",
-    "webpack-dev-server": "^2.11.1"
-  },
+  "name": "mappa",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "dotenv-webpack": "^1.5.4",
-    "handlebars": "^4.0.11",
-    "handlebars-loader": "^1.6.0",
-    "jquery": "^3.3.1",
-    "knockout": "^3.5.0-beta",
-    "webpack": "^3.10.0"
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-google-maps": "^9.4.5",
+    "react-scripts": "1.1.1",
+    "serve": "^6.5.0"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "now-start": "serve --single ./build"
   }
 }
