--- webpack.config.js
@@ -1,48 +1,0 @@
-'use strict';
-
-const path = require('path');
-const webpack = require('webpack');
-const Dotenv = require('dotenv-webpack');
-
-module.exports = {
-    resolve: {
-        alias: {
-            handlebars: 'handlebars/dist/handlebars.min.js'
-        }
-    },
-    entry: './src/app.js',
-    output: {
-        path: path.resolve(__dirname, 'dist'),
-        filename: 'bundle.js',
-        publicPath: '/dist'
-    },
-    plugins: [
-        new webpack.ProvidePlugin({
-            $: 'jQuery',
-            jQuery: 'jQuery'
-        }),
-        new Dotenv({
-            path: '.env',
-            safe: false
-        }),
-    ],
-     module: {
-        noParse: /node_modules\/knockout\/build\/output\/*.js/,
-        rules: [
-            {
-                test: /\.js$/,
-                use: [
-                    {
-                        loader: 'babel-loader',
-                        options: {
-                            presets: ['env']
-                        }
-                    }
-                ]
-            },
-            { 
-                test: /\.handlebars$/, 
-                loader: "handlebars-loader" 
-            }
-    ]
-}}
